Flask Api/requirements.txt
Flask==1.1.1
Flask-RESTful==0.3.7
mysql-connector-python==8.0.18
Flask-Docs

